// assets/js/types.ts
export interface ScrollToTopConfig {
    readonly buttonD: string;
    readonly buttonT: string;
    readonly shadowSize: string;
    readonly roundnessSize: string;
    readonly buttonDToBottom: string;
    readonly buttonDToRight: string;
    readonly selectedBackgroundColor: string;
    readonly selectedIconColor: string;
    readonly buttonWidth: string;
    readonly buttonHeight: string;
    readonly svgWidth: string;
    readonly svgHeight: string;
}
// ...
export type Nullable<T> = T | null;